requirements.txt
PyJWT
Click==7.0
codeclimate-test-reporter==0.2.3
coverage==4.3
Flask==1.0.2
Flask-API==1.0
Flask-Classful==0.14.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.16.0
python-dotenv==0.9.1
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
